requirements.txt
fastapi
uvicorn
pydantic-ai
pypdf2
python-dotenv
python-multipart